changes httpclient to propogate the http status code
If the node replies with a non ok status code and the json response
can not be parsed, an ErrNonOKStatusCodeFromAPIRequest is returned.
If the response could be parsed and the status code is not ok,
the error or exception from the payload is extracted.
If the payload is a JSON response but doesn't include either
an error or exception field, ErrNonErrorResponse is returned.

Signed-off-by: Luca Moser <moser.luca@gmail.com>